Update the blogSlice file
