+++ app/javascript/components/Blogs/EditBlogForm.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import { selectBlogById } from '../redux/blogSlice';
+import { selectBlogById } from '../../redux/blogSlice';
 
-const SingleBlogPage = ({ match }) => {
+const EditBlogForm = ({ match }) => {
   const { blogId } = match.params;
 
   const blog = useSelector((state) => selectBlogById(state, blogId));
 
-  return <div>SingleBlogPage</div>;
+  return <div>EditBlogForm</div>;
 };
 
-export default SingleBlogPage;
+export default EditBlogForm;
