+++ app/javascript/components/Blogs/SingleBlogPage.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import { selectBlogById } from '../redux/blogSlice';
+import { selectBlogById } from '../../redux/blogSlice';
 
-const EditBlogForm = ({ match }) => {
+const SingleBlogPage = ({ match }) => {
   const { blogId } = match.params;
 
   const blog = useSelector((state) => selectBlogById(state, blogId));
 
-  return <div>EditBlogForm</div>;
+  return <div>SingleBlogPage</div>;
 };
 
-export default EditBlogForm;
+export default SingleBlogPage;
